mobile-app/server/requirements.txt
opencv_python==4.5.4.60
firebase_admin==4.5.1
Flask==1.1.2
numpy==1.18.4
Pillow==9.1.0
protobuf==3.20.1
